@@ -1,20 +1,14 @@
 source 'https://rubygems.org'
 
-ruby '2.4.5'
-gem 'rails', '4.2.11'
-
-# gem 'rack', '>= 1.6.11'
-# gem 'ffi', '>= 1.9.24'
-# gem 'sprockets', '>= 3.7.2'
-# gem 'rails-html-sanitizer', '>= 1.0.4'
-# gem 'loofah', '>= 2.2.3'
-
-gem 'sass-rails', '~> 5.0'
-gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.1.0'
+ruby '2.7.2'
+gem 'rails', '5.2.4.5'
+gem 'sass-rails', '5.1.0'
+gem 'uglifier', '1.3.0'
+gem 'coffee-rails', '5.0.0'
 gem 'jquery-rails'
 gem 'turbolinks'
-gem 'omniauth-google-oauth2', '~> 0.5.3'
+gem 'omniauth', '1.9.1'
+gem 'omniauth-google-oauth2', '0.7.0'
 gem 'devise'
 gem 'thin'
 gem 'figaro'
